修改了 train.cu 中的 sqrt 函数调用，将它们变为 sqrtf (#82)
有些时候编译器通不过这个sqrt的函数调用，这个pr将修复上述问题

+++ ppq/csrc/cuda/train.cu
@@ -137,7 +137,7 @@ __global__ void _RoundingLoss_LT(
         diff_sum += diff;
     }
     float diff_reduced = BlockReduceSum<float>(diff_sum);
-    if(threadIdx.x == 0) atomicAdd(out, diff_reduced / sqrt(num_of_element));
+    if(threadIdx.x == 0) atomicAdd(out, diff_reduced / sqrtf(num_of_element));
 }
 
 __host__ Tensor RoundingLoss_LT(
@@ -189,7 +189,7 @@ __global__ void _RoundingLoss_LT_B(
         // if value has been clipped, set grad = 0
         if (v > s * (clip_max - o)) grad = 0;
         if (v < s * (clip_min - o)) grad = 0;
-        dx[iter] = grad / sqrt(num_of_element);
+        dx[iter] = grad / sqrtf(num_of_element);
     }
 }
 
@@ -243,7 +243,7 @@ __global__ void _RoundingLoss_LC(
         diff_sum += diff;
     }
     float diff_reduced = BlockReduceSum<float>(diff_sum);
-    if(threadIdx.x == 0) atomicAdd(out, diff_reduced / sqrt(num_of_element));
+    if(threadIdx.x == 0) atomicAdd(out, diff_reduced / sqrtf(num_of_element));
 }
 
 __host__ Tensor RoundingLoss_LC(
@@ -254,7 +254,7 @@ __host__ Tensor RoundingLoss_LC(
     /*
         Compute rounding loss R of a given tensor T.
 
-        R = sum(norm2(T - quant(T))) / sqrt(elements_of(T))
+        R = sum(norm2(T - quant(T))) / sqrtf(elements_of(T))
 
         If one value inside R is clipped by quant function, 
             then it contributes nothing to R.
@@ -303,7 +303,7 @@ __global__ void _RoundingLoss_LC_B(
         // if value has been clipped, set grad = 0
         if (v > scale[c] * (clip_max - offset[c])) grad = 0;
         if (v < scale[c] * (clip_min - offset[c])) grad = 0;
-        dx[iter] = grad / sqrt(num_of_element);
+        dx[iter] = grad / sqrtf(num_of_element);
     }
 }
